Fix bug FM-836 and FM-823
For FM-836, there is a problem with HTTPNetStream when an external buffer manager changes the bufferTime of the stream to be larger than 3.75 for fast bandwidth and 7.5 for slow bandwidth. When OSMFPlayer runs into buffer empty event, the buffer manager will increment the bufferTime. After a few empty buffer events, the bufferTime will be larger than 3.75 or 7.5, at which point the next buffer empty event will put FP into the buffering state but never comes out because HTTPNetStream will stop downloading fragments and send the tc messages to appendBytes. The fix takes the bufferTime into consideration and choose the max of buffer time and 3.75/7.5 as the threshold.

For FM-823, we set the default to enable enhanced seek.

git-svn-id: http://svn.code.sf.net/adobe/osmf/svn/osmf/trunk@1680 a9308255-753e-0410-a2e9-80b3fbc4fff6