Fix for FM-1295 (http://bugs.adobe.com/jira/browse/FM-1295)
NOTE
When dealing with protected content, the drm side car information was loaded only from drmContentData property of StreamingURLResource. This was particularly bad for MBR content, where the drm information was loaded to late and could put the MediaElement in an unloaded and back in a loaded state.

Now we are looking for drm information in the following order
- in drmContentData 
- in DRM namespace on the resource ( here we are trying to obtain the drm metadata for the stream indicated by initial index, and if we don't succeed then we are using an available information )



git-svn-id: http://svn.code.sf.net/adobe/osmf/svn/osmf/trunk@2108 a9308255-753e-0410-a2e9-80b3fbc4fff6